+++ src/Console/ApplicationFactory.php
@@ -3,16 +3,14 @@
 namespace Castor\Console;
 
 use Castor\Console\Command\CompileCommand;
+use Castor\Console\Command\ComposerCommand;
 use Castor\Console\Command\DebugCommand;
 use Castor\Console\Command\RepackCommand;
 use Castor\Container;
-use Castor\Descriptor\DescriptorsCollection;
-use Castor\Event\AfterApplicationInitializationEvent;
 use Castor\Helper\PathHelper;
 use Castor\Helper\PlatformHelper;
 use Castor\Monolog\Processor\ProcessProcessor;
-use Joli\JoliNotif\Notifier;
-use Joli\JoliNotif\NotifierFactory;
+use Joli\JoliNotif\DefaultNotifier;
 use Monolog\Logger;
 use Psr\Cache\CacheItemPoolInterface;
 use Psr\Container\ContainerInterface;
@@ -40,6 +38,7 @@ use Symfony\Component\HttpClient\HttpClient;
 use Symfony\Component\String\Slugger\AsciiSlugger;
 use Symfony\Component\VarDumper\Cloner\AbstractCloner;
 use Symfony\Contracts\Cache\CacheInterface;
+use Symfony\Contracts\EventDispatcher\Event;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 use function Symfony\Component\DependencyInjection\Loader\Configurator\service;
@@ -84,9 +83,8 @@ class ApplicationFactory
     {
         $errorHandler = ErrorHandler::register();
 
-        AbstractCloner::$defaultCasters[self::class] = ['Symfony\Component\VarDumper\Caster\StubCaster', 'cutInternals'];
-        AbstractCloner::$defaultCasters[AfterApplicationInitializationEvent::class] = ['Symfony\Component\VarDumper\Caster\StubCaster', 'cutInternals'];
-        AbstractCloner::$defaultCasters[DescriptorsCollection::class] = ['Symfony\Component\VarDumper\Caster\StubCaster', 'cutInternals'];
+        AbstractCloner::$defaultCasters[Application::class] = ['Symfony\Component\VarDumper\Caster\StubCaster', 'cutInternals'];
+        AbstractCloner::$defaultCasters[Event::class] = ['Symfony\Component\VarDumper\Caster\StubCaster', 'cutInternals'];
 
         return $errorHandler;
     }
@@ -172,8 +170,7 @@ class ApplicationFactory
 
             ->set(AsciiSlugger::class)
 
-            ->set(Notifier::class)
-                ->factory([NotifierFactory::class, 'create'])
+            ->set(DefaultNotifier::class)
 
             ->set(Container::class)
                 ->public()
@@ -204,6 +201,7 @@ class ApplicationFactory
         ;
         if (!$repacked) {
             $app
+                ->call('add', [service(ComposerCommand::class)])
                 ->call('add', [service(RepackCommand::class)])
                 ->call('add', [service(CompileCommand::class)])
             ;