+++ src/Helper/Notifier.php
@@ -2,28 +2,37 @@
 
 namespace Castor\Helper;
 
+use Castor\ContextRegistry;
+use Joli\JoliNotif\DefaultNotifier;
 use Joli\JoliNotif\Exception\InvalidNotificationException;
 use Joli\JoliNotif\Notification;
-use Joli\JoliNotif\Notifier as JoliNotifier;
-use Joli\JoliNotif\Notifier\NullNotifier;
 use Psr\Log\LoggerInterface;
+
+use function Castor\context;
 
 class Notifier
 {
     public function __construct(
-        private JoliNotifier $notifier,
+        private DefaultNotifier $notifier,
         private LoggerInterface $logger,
+        private ContextRegistry $contextRegistry
     ) {
     }
 
-    public function send(string $message): void
+    public function send(string $message, ?string $title = null): void
     {
+        if (false === $this->contextRegistry->getCurrentContext()->notify) {
+            return;
+        }
+
         $notification = (new Notification())
-            ->setTitle('Castor')
+            ->setTitle($title ?? $this->getNotifyTitle())
             ->setBody($message)
         ;
 
-        if ($this->notifier instanceof NullNotifier) {
+        $driver = $this->notifier->getDriver();
+
+        if (null === $driver) {
             $this->logger->warning('No supported notifier found, notification not sent.');
 
             return;
@@ -40,5 +49,14 @@ class Notifier
         } catch (\Throwable $e) {
             $this->logger->error('Failed to send notification: ' . $e->getMessage());
         }
+    }
+
+    private function getNotifyTitle(): string
+    {
+        if ('' !== context()->notificationTitle) {
+            return context()->notificationTitle;
+        }
+
+        return 'Castor';
     }
 }