+++ src/Import/Mount.php
@@ -12,6 +12,8 @@ class Mount
         public readonly string $path,
         public readonly bool $allowEmptyEntrypoint = false,
         public readonly ?string $namespacePrefix = null,
+        public readonly bool $allowRemotePackage = true,
+        public readonly ?string $file = null,
     ) {
     }
 }