+++ tests/fixtures/broken/import-invalid-package/castor.php
@@ -2,4 +2,4 @@
 
 use function Castor\import;
 
-import('package://foo/bar', source: ['type' => 'git']);
+import('package://foo/bar');